+++ src/engine/enginestartdialog.h
@@ -113,7 +113,7 @@ private slots:
     void reListNamespacesError(const QString &error);
     void reVersion(const QString &engineVersion, const QString &gamsVersion, bool inKubernetes);
     void reVersionError(const QString &errorText);
-    void reUserInstances(const QList<QPair<QString, QList<int> > > instances, const QString &defaultLabel);
+    void reUserInstances(const QList<QPair<QString, QList<double> > > instances, const QString &defaultLabel);
     void reUserInstancesError(const QString &errorText);
     void quotaHint(const QStringList &diskHint, const QStringList &volumeHint);
     void forceGdxStateChanged(int state);
